refactor(agent): 抽出 root-agent 扩展并加固 getSnapshot 隔离
- 新增 extensions/ 目录与 RootAgentExtensionHost 端口接口，把
  WakeReminder / ContextCompaction / SnapshotPersistence / LlmTelemetry /
  ToolExecutionState / ToolFallback / PostToolEffects 七个扩展从
  root-agent-runtime.ts 中独立出来；ReAct kernel extensions 现在依赖
  Pick<RootAgentExtensionHost, ...> 而非整个 host 类。host 对扩展的契约
  显式化，KV 缓存安全 invariant 写在端口接口注释里。
  root-agent-runtime.ts 1188 行 -> 990 行。
- DefaultAgentContext.getSnapshot 在出口处 cloneMessages，调用方拿到的
  是私有副本，避免下游意外修改污染稳定前缀；fork / exportPersistedSnapshot
  里冗余的二次 clone 同步去掉。
